adds tests for editing status
